GCA-167 Implement Workload microservice - Create service layer

+++ trainers-workload-service/src/main/java/com/trainersworkloadservice/model/dto/MonthlyWorkloadRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 
 public record MonthlyWorkloadRequest(
         @NotBlank String username,
-        @NotBlank Short month,
-        @NotBlank Short year
+        @NotBlank int month,
+        @NotBlank int year
 ) {
 }
