+++ trainers-workload-service/src/main/java/com/trainersworkloadservice/model/dto/MonthlyWorkloadResponse.java
@@ -2,9 +2,9 @@ package com.trainersworkloadservice.model.dto;
 
 public record MonthlyWorkloadResponse(
         String username,
-        Short year,
-        Short month,
-        Long totalHours,
-        Long totalMinutes
+        int year,
+        int month,
+        long totalHours,
+        long totalMinutes
 ) {
 }
